perf(mcp-settings): isolate log tab updates
